Use AlignedAlloc for ThreadLocal
Windows is complaining that new[] is using _aligned_alloc and delete[] is using free. So we just make it explicit
